Adjust Seldon example
