+++ examples/seldon_deployment/steps/common/importer.py
@@ -1,0 +1,58 @@
+#  Copyright (c) ZenML GmbH 2023. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at:
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+#  or implied. See the License for the specific language governing
+#  permissions and limitations under the License.
+
+
+import numpy as np  # type: ignore [import]
+import pandas as pd
+import requests  # type: ignore [import]
+import tensorflow as tf  # type: ignore [import]
+
+from zenml import step
+from zenml.steps import Output
+
+
+@step
+def importer_mnist() -> (
+    Output(
+        x_train=np.ndarray,
+        y_train=np.ndarray,
+        x_test=np.ndarray,
+        y_test=np.ndarray,
+    )
+):
+    """Download the MNIST data store it as an artifact."""
+    (x_train, y_train), (
+        x_test,
+        y_test,
+    ) = tf.keras.datasets.mnist.load_data()
+    return x_train, y_train, x_test, y_test
+
+
+def get_data_from_api():
+    url = (
+        "https://storage.googleapis.com/zenml-public-bucket/mnist"
+        "/mnist_handwritten_test.json"
+    )
+
+    df = pd.DataFrame(requests.get(url).json())
+    data = df["image"].map(lambda x: np.array(x)).values
+    data = np.array([x.reshape(28, 28) for x in data])
+    return data
+
+
+@step(enable_cache=False)
+def dynamic_importer() -> Output(data=np.ndarray):
+    """Downloads the latest data from a mock API."""
+    data = get_data_from_api()
+    return data
