+++ examples/seldon_deployment/steps/sklearn/evaluator.py
@@ -1,0 +1,30 @@
+#  Copyright (c) ZenML GmbH 2023. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at:
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+#  or implied. See the License for the specific language governing
+#  permissions and limitations under the License.
+
+
+import numpy as np  # type: ignore [import]
+from sklearn.base import ClassifierMixin
+
+from zenml import step
+
+
+@step
+def sklearn_evaluator(
+    x_test: np.ndarray,
+    y_test: np.ndarray,
+    model: ClassifierMixin,
+) -> float:
+    """Calculate accuracy score with classifier."""
+    test_acc = model.score(x_test.reshape((x_test.shape[0], -1)), y_test)
+    return test_acc
