+++ examples/seldon_deployment/steps/sklearn/preprocessor.py
@@ -1,0 +1,29 @@
+#  Copyright (c) ZenML GmbH 2023. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at:
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+#  or implied. See the License for the specific language governing
+#  permissions and limitations under the License.
+
+import json
+
+import numpy as np  # type: ignore [import]
+
+from zenml import step
+from zenml.steps import Output
+
+
+@step
+def sklearn_predict_preprocessor(input: np.ndarray) -> Output(data=str):
+    """Prepares the data for inference."""
+    input = input / 255.0
+    input = input.reshape((input.shape[0], -1))
+    request = input.tolist()
+    return json.dumps(request)
